chore: silence clippy 1.95 lint regressions across workspace
Rust 1.95.0 promoted several lints to default-warn (and CI runs with
`-D warnings`), causing a cascade of build failures across mlxcel-core
and mlxcel-surgery that had nothing to do with their underlying logic.
This commit makes the workspace lint-clean again under the new toolchain.

Auto-fixed via `cargo clippy --fix`:

- `manual_is_multiple_of` → `.is_multiple_of()` (pack3, quant3, utils)
- `manual_div_ceil` → `.div_ceil()` (utils)
- `unnecessary_cast` (quant, codebook, generate, etc.)
- `useless_conversion` (generate)
- `double_parens` (mtp/tests)

Manually rewritten:

- `needless_range_loop` → iterator + take/enumerate where the conversion
  was clear (utils, quant, quant3, sampling, boundary, turbo_tests).
- `field_reassign_with_default` → struct-init form (generate, speculative).
- `doc_overindented_list_items` → list-continuation indent normalised
  (speculative/mtp/generator).
- `doc_lazy_continuation` → reworded the `softmax + cold-V` paragraph in
  turbo_tests so a leading `+` no longer parses as a markdown list marker.
- `unusual_byte_groupings` → `0xC0FF_EE42` (pack3, quant).
- `duplicated_attribute` → drop inner `#![cfg(test)]` where the parent
  module already gates with `#[cfg(test)]` (mlxcel-surgery).
- Added a `# Safety` section to `attention_from_ptr` (layers).

Targeted `#[allow(...)]` (intentional shape, not worth restructuring):

- `too_many_arguments` on the attention-dispatch helpers in `layers.rs`
  and the `attention_sparse_v_turbo4{,_fused}` kernels: positional
  signatures intentionally mirror Metal/MLX call sites.
- `needless_range_loop` on `forward_batched`: mixed indexing into
  `input_ids` slices and `batch_caches[i]` is clearer than enumerate.
- `large_enum_variant` / `result_large_err` on the paged-cache adopt
  path: shrinking the variant would force callers through `Box` for no
  runtime gain on a cold path.
- `type_complexity` on the `SyntheticTarget` test fixture's
  `rollback_events`: it is a one-off test type.

`cargo clippy --workspace --all-targets -- -D warnings` is clean and
`cargo test --workspace` matches the pre-existing baseline (the same two
`multimodal::video` fd-variant tests fail on `main` too, due to runtime
ffmpeg policy mismatches unrelated to this change).

+++ src/lib/mlxcel-core/src/speculative/mtp/tests.rs
@@ -579,12 +579,7 @@ fn greedy_parity_perfect_drafter_matches_no_drafter_baseline_32_tokens() {
     let scripted_target: Vec<Vec<i32>> = (0..11)
         .map(|r| {
             let base = 1000 + r * 4;
-            vec![
-                base as i32,
-                (base + 1) as i32,
-                (base + 2) as i32,
-                (base + 3) as i32,
-            ]
+            vec![base, (base + 1), (base + 2), (base + 3)]
         })
         .collect();
     let scripted_draft: Vec<Vec<i32>> = scripted_target
